requirements.txt
adversarial-robustness-toolbox==1.19.0
tensorflow==2.14.0
torch==1.13.1
numpy==1.26.0
scipy==1.10.1
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.4.1.post1
keras==2.14.0
opencv-python==4.7.0.68
pandas==2.2.3
h5py==3.10.0
tqdm==4.67.1
Pillow==11.1.0
